Finished upgrade buttons and adapated for react native Updated API function for react native Updated timer for react native Added cookie images
