+++ app/index.tsx
@@ -2,10 +2,12 @@ import { useState, useEffect } from "react";
 import { Text, View } from "react-native";
 
 import Shop1 from "./Upgrades/Shop1";
-import CookieButton from "./"
-
+import Shop2 from "./Upgrades/Shop2";
+import CookieButton from "./";
 
 export default function Index() {
+  // Consts
+
   const [items, setItems] = useState<any[]>([]);
   const [count, setCount] = useState(0);
   const [shop1, setShop1] = useState(0);
@@ -14,41 +16,24 @@ export default function Index() {
   const [shop2Count, setShop2Count] = useState(0);
   const [cookieData, setCookieData] = useState(0);
 
+  // Consts
+
   // API
 
-  // useEffect(() => {
-  //   async function fetchData() {
-  //     try {
-  //       const response = await fetch(
-  //         "https://cookie-upgrade-api.vercel.app/api/upgrades"
-  //       );
-  //       const data = await response.json();
-  //       console.log(`Data Received`);
-  //       setItems(data);
-  //       setShop1Count(data[shop1].increase);
-  //       setShop2Count(data[shop2 - 1].increase);
-  //       setCookieData(data);
-  //     } catch (error) {
-  //       console.error("Data failed");
-  //     }
-  //   }
-  //   fetchData();
-  // }, [items, shop1, shop2, shop1Count, shop2Count]);
-
-  async function fetchData() {
-    const response = await fetch(
-      "https://cookie-upgrade-api.vercel.app/api/upgrades"
-    );
-    const data = await response.json();
-    console.log(`Data Received`);
-    setItems(data);
-    setShop1Count(data[shop1].increase);
-    setShop2Count(data[shop2].increase);
-    setCookieData(data);
-  }
-  fetchData();
-
-  // console.log(items);
+  useEffect(() => {
+    async function fetchData() {
+      const response = await fetch(
+        "https://cookie-upgrade-api.vercel.app/api/upgrades"
+      );
+      const data = await response.json();
+      console.log(`Data Received`);
+      setItems(data);
+      setShop1Count(data[shop1].increase);
+      setShop2Count(data[shop2].increase);
+      setCookieData(data);
+    }
+    fetchData();
+  }, [shop1, shop2, items]);
 
   // API
 
@@ -65,8 +50,6 @@ export default function Index() {
 
   // Timer
 
-  // Upgrade Button
-
   return (
     <View
       style={{
@@ -75,30 +58,51 @@ export default function Index() {
         alignItems: "center",
       }}
     >
-      <Text>{count}</Text>
-      <div id="Shop1Button">
-        {items.map((item) => {
-          return (
-            item.id == shop1 + 1 && (
-              <Shop1
-                key={item.id}
-                heading={item.name}
-                shopValue={shop1}
-                setShop={setShop1}
-                setCount={setCount}
-                shopID={item.id}
-                count={count}
-                upgradeCost={item.cost}
-                clicks={() => setShop1(item.id)}
-              />
-            )
-          );
-        })}
-      </div>
-      <button id="Shop2Button">Shop2</button>
+      <Text>Count:{count}</Text>
+
+      {/*  Upgrade Buttons */}
+
+      {/* <div id="Shop1Button"> */}
+      {items.map((item) => {
+        return (
+          item.id == shop1 + 1 && (
+            <Shop1
+              key={item.id}
+              heading={item.name}
+              shopValue={shop1}
+              setShop={setShop1}
+              setCount={setCount}
+              shopID={item.id}
+              count={count}
+              upgradeCost={item.cost}
+              clicks={() => setShop1(item.id)}
+            />
+          )
+        );
+      })}
+      {items.map((item) => {
+        return (
+          item.id == shop2 + 1 && (
+            <Shop2
+              key={item.id}
+              heading={item.name}
+              shopValue={shop2}
+              setShop={setShop2}
+              setCount={setCount}
+              shopID={item.id}
+              count={count}
+              upgradeCost={item.cost}
+              clicks={() => setShop2(item.id)}
+            />
+          )
+        );
+      })}
+      {/* </div> */}
+
+      {/* Upgrade Buttons */}
+
       <Text>Edit app/index.tsx to edit this screen.</Text>
       <CookieButton />
-
     </View>
   );
 }
